refactor: Move exported values to their own constants file
