+++ packages/avatar/src/index.js
@@ -1,7 +1,4 @@
 export { default } from "./Avatar";
 export { sizes, AVAILABLE_SIZES } from "./sizes";
-export {
-  default as stylesheet,
-  SizeMapping,
-  StyleItems
-} from "./Avatar.stylesheet";
+export { default as stylesheet } from "./Avatar.stylesheet";
+export * from "./constants";
